% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleRBPs.R
\name{BubbleRBPs}
\alias{BubbleRBPs}
\title{BubbleRBPs}
\usage{
BubbleRBPs(res = res, clusters = clusters, rbp_lfc = rbp_lfc)
}
\arguments{
\item{symbol}{RBP_res, clusters, rbp_lfc}
}
\value{
NULL. Generates a bubble plot
}
\description{
Use clusters from DeltaTE output to generate enrichement plot of RBPs
}
\examples{
BubbleRBPs(RBP_res=res,clusters=clusters,rbp_lfc=rbp_lfc)

}
