% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIPreg.R
\name{CLIPreg}
\alias{CLIPreg}
\title{CLIPreg}
\usage{
CLIPreg(RBP_data = RBP_POSTAR, gene_groups = gene_groups)
}
\arguments{
\item{symbol}{folder, RBP_data, gene_groups}
}
\value{
List of gene_groups containing dataframe with all the enrichment score for each RBP
}
\description{
Calculates the enrichment of each RBP for each gene group
}
\examples{
CLIPreg(RBP_data=RBP_POSTAR,gene_groups=gene_groups)
}
