% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIPReg_V3.R
\name{CLIPReg_V3}
\alias{CLIPReg_V3}
\title{CLIPReg}
\usage{
CLIPReg_V3(
  folder = "Folder",
  RBP_data = "rbp_gene_postar.txt",
  background = "bg.txt",
  cluster = "Tx_down.txt"
)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
