% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIPReg_V4.R
\name{CLIPReg_V4}
\alias{CLIPReg_V4}
\title{CLIPReg_V3}
\usage{
CLIPReg_V4(RBP_data = RBP_POSTAR, clusters = clusters)
}
\arguments{
\item{symbol}{folder, RBP_data, clusters}
}
\value{
List of clusters containing dataframe with all the enrichment score for each RBP
}
\description{
Calculates the enrichment of each RBP for each cluster
}
\examples{
CLIPreg_V4(RBP_data=RBP_POSTAR,clusters=clusters)

}
