% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{combine}
\alias{combine}
\title{combine}
\usage{
combine(res1 = res_Encode, res2 = res_Postar, FDR = 0.1)
}
\arguments{
\item{symbol}{gene_groups_file}
}
\description{
Load the gene_groups file from user
}
