% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{combine_targets}
\alias{combine_targets}
\title{combine_targets}
\usage{
combine_targets(
  RBP_list1 = RBP_ENCODE,
  RBP_list2 = RBP_POSTAR,
  background = gene_groups$geneID
)
}
\arguments{
\item{symbol}{RBP_list1, RBP_list2, background}
}
\description{
Combine targets from POSTAR and ENCODE and keep only genes from background
}
