% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_network_by_group.R
\name{Draw_network_by_group}
\alias{Draw_network_by_group}
\title{Draw_network_by_group}
\usage{
Draw_network_by_group(
  rbp_lfc = rbp_lfc,
  res = res,
  Targets = Targets,
  clusters = clusters,
  n = 5,
  forwarded = F
)
}
\arguments{
\item{symbol}{rbp_lfc, res, Targets, clusters, n}
}
\value{
Plot network
}
\description{
This function draw a network from CLIPreg output. It plots the n most changing RBP and their targets group by TE category
}
\examples{
Draw_network(rbp_lfc=rbp_lfc,res=res_both,clusters=clusters,n=5)

}
