% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTarget.R
\name{getTarget}
\alias{getTarget}
\title{CLIPReg}
\usage{
getTarget(
  folder = "Folder",
  RBP_data = "rbp_gene_postar.txt",
  background = "bg.txt"
)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
