% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapRBP.R
\name{HeatmapRBP}
\alias{HeatmapRBP}
\title{CLIPReg}
\usage{
HeatmapRBP(res = res, RBP_change = rbp_lfc)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
