% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapTargets.R
\name{HeatmapTargets}
\alias{HeatmapTargets}
\title{CLIPReg}
\usage{
HeatmapTargets(
  folder = "Folder",
  Clusters = clusters,
  RBP_name = "RBM47",
  targets = Targets,
  counts_rna = counts_rna,
  counts_ribo = counts_ribo,
  scale = T,
  Not_DE = F
)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
