% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_clusters}
\alias{load_clusters}
\title{load_clusters}
\usage{
load_clusters(cluster_file = cluster_file)
}
\arguments{
\item{symbol}{cluster_file}
}
\description{
Load the cluster file from user
}
