% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_gene_groups}
\alias{load_gene_groups}
\title{load_gene_groups}
\usage{
load_gene_groups(gene_groups_file = gene_groups_file)
}
\arguments{
\item{symbol}{gene_groups_file}
}
\description{
Load the gene group file from user
}
