% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_ribo_lfc}
\alias{load_ribo_lfc}
\title{load_ribo_lfc}
\usage{
load_ribo_lfc(ribo_lfc_file = ribo_lfc_file)
}
\arguments{
\item{symbol}{ribo_lfc}
}
\description{
Load the ribo_lfc file from user
}
