% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_ribo_tpm}
\alias{load_ribo_tpm}
\title{load_ribo_tpm}
\usage{
load_ribo_tpm(ribo_tpm_file = ribo_tpm_file)
}
\arguments{
\item{symbol}{ribo_tpm}
}
\description{
Load the ribo_tpm file from user
}
