% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GO.R
\name{Plot_GO}
\alias{Plot_GO}
\title{Plot_GO}
\usage{
Plot_GO(
  rbp_lfc = rbp_lfc,
  res = res,
  Targets = Targets,
  gene_groups = gene_groups,
  n = 5,
  tpm_ribo = tpm_ribo,
  th = 200,
  GO_to_show = 3,
  forwarded = F
)
}
\arguments{
\item{symbol}{}
}
\description{
This function draws a network from CLIPreg output. Precise the number of RBP to consider with n
the thershold for size of node you want the GO done with th. The number of GO term to show per node with GO_to_show
}
\examples{

}
