% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GO_node_name.R
\name{Plot_GO_node_name}
\alias{Plot_GO_node_name}
\title{Plot_GO}
\usage{
Plot_GO_node_name(
  rbp_lfc = rbp_lfc,
  res = res_both[-c(1, 2)],
  Targets = Targets,
  clusters = clusters,
  n = 5,
  all_genes = rownames(tpm_ribo),
  Nodes_to_keep = c(19, 15),
  GO_to_show = 3
)
}
\arguments{
\item{symbol}{}
}
\description{
This function draws a network from CLIPreg output. Precise the number of RBP to consider with n
the thershold for size of node you want the GO done with th. The number of GO term to show per node with GO_to_show
}
\examples{

}
