% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GO_RBP.R
\name{Plot_GO_RBP}
\alias{Plot_GO_RBP}
\title{Plot_GO}
\usage{
Plot_GO_RBP(
  rbp_of_interest = "QKI",
  tpm_ribo = tpm_ribo,
  Targets = Targets,
  gene_groups = gene_groups,
  GO_to_show = 3
)
}
\arguments{
\item{symbol}{}
}
\description{
This function draws a network from CLIPreg output. Precise the number of RBP to consider with n
the thershold for size of node you want the GO done with th. The number of GO term to show per node with GO_to_show
}
\examples{

}
