% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{rbp_change}
\alias{rbp_change}
\title{rbp_change}
\usage{
rbp_change(res = res, ribo_lfc = ribo_lfc)
}
\arguments{
\item{symbol}{res, ribo_lfc}
}
\description{
Subset ribo_lfc to keep only rbp_lfc
}
