% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_heatmap.R
\name{save_heatmap}
\alias{save_heatmap}
\title{CLIPReg}
\usage{
save_heatmap(res = res, RBP_change = rbp_lfc, location = location)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
