% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleRBPs.R
\name{BubbleRBPs}
\alias{BubbleRBPs}
\title{CLIPReg}
\usage{
BubbleRBPs(RBP_res = res, clusters = clusters)
}
\arguments{
\item{symbol}{}
}
\description{
This package identifies...
}
\examples{
RbpReg('AAPL')

}
