% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_network.R
\name{Draw_network}
\alias{Draw_network}
\title{Draw_network}
\usage{
Draw_network(rbp_lfc = rbp_lfc, res = res_both, clusters = clusters, n = 5)
}
\arguments{
\item{symbol}{rbp_lfc, res, clusters, n}
}
\value{
Plot network
}
\description{
This function draw a network from CLIPreg output. It plots the n most changing RBP and their target
}
\examples{
Draw_network(rbp_lfc=rbp_lfc,res=res_both,clusters=clusters,n=5)

}
