% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTarget.R
\name{GetTarget}
\alias{GetTarget}
\title{GetTarget}
\usage{
GetTarget(RBP_data = RBP_POSTAR, background = clusters$geneID)
}
\arguments{
\item{symbol}{folder, RBP_data, background}
}
\value{
List of RBPs each containing 1 vector of targets in form of characters
}
\description{
Summarize RBPs in a list with their targets
}
\examples{
GetTarget(RBP_data=RBP_POSTAR, background=clusters$geneID)

}
