% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapRBP.R
\name{HeatmapRBP}
\alias{HeatmapRBP}
\title{HeatmapRBP}
\usage{
HeatmapRBP(res = res, rbp_lfc = rbp_lfc)
}
\arguments{
\item{symbol}{res, rbp_lfc}
}
\value{
Plot a heatmap
}
\description{
Plot a heatmap of the pvalues of enrichment of each RBP in each cluster
}
\examples{
HeatmapRBP(res=res,RBP_change=rbp_lfc)

}
