% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapTargets.R
\name{HeatmapTargets}
\alias{HeatmapTargets}
\title{HeatmapTargets}
\usage{
HeatmapTargets(
  folder = "Folder",
  Clusters = clusters,
  RBP_name = "RBM47",
  targets = Targets,
  counts_rna = counts_rna,
  counts_ribo = counts_ribo,
  scale = T,
  Not_DE = F
)
}
\arguments{
\item{symbol}{folder, Clusters, RBP_name, targets, counts_rna, counts_ribo, scale, Not_DE)}
}
\value{
Plot a heatmap
}
\description{
Plot the counts of the targets of a specified RBP
}
\examples{
HeatmapTargets(folder="Folder",Clusters=clusters,RBP_name="RBM47",targets=Targets,counts_rna=counts_rna, counts_ribo=counts_ribo,scale=T,Not_DE=F)

}
